Move yae::SliderDrag to yaeScrollview.h
