+++ apprenticevideo/yaeScrollview.cpp
@@ -201,4 +201,53 @@ namespace yae
   }
 #endif
 
+
+  //----------------------------------------------------------------
+  // SliderDrag::SliderDrag
+  //
+  SliderDrag::SliderDrag(const char * id,
+                         const Canvas::ILayer & canvasLayer,
+                         Scrollview & scrollview,
+                         Item & scrollbar):
+    InputArea(id),
+    canvasLayer_(canvasLayer),
+    scrollview_(scrollview),
+    scrollbar_(scrollbar),
+    startPos_(0.0)
+  {}
+
+  //----------------------------------------------------------------
+  // SliderDrag::onPress
+  //
+  bool
+  SliderDrag::onPress(const TVec2D & itemCSysOrigin,
+                      const TVec2D & rootCSysPoint)
+  {
+    startPos_ = scrollview_.position_;
+    return true;
+  }
+
+  //----------------------------------------------------------------
+  // SliderDrag::onDrag
+  //
+  bool
+  SliderDrag::onDrag(const TVec2D & itemCSysOrigin,
+                     const TVec2D & rootCSysDragStart,
+                     const TVec2D & rootCSysDragEnd)
+  {
+    double bh = scrollbar_.height();
+    double sh = this->height();
+    double yRange = bh - sh;
+
+    double dy = rootCSysDragEnd.y() - rootCSysDragStart.y();
+    double ds = dy / yRange;
+    double t = std::min<double>(1.0, std::max<double>(0.0, startPos_ + ds));
+    scrollview_.position_ = t;
+
+    parent_->uncache();
+    canvasLayer_.delegate()->requestRepaint();
+
+    return true;
+  }
+
 }
